Added GitHub UserName/ Added the seller story in the story document
I added the github username for each member as it is part of the grading scheme. We also need to write the role/strengths of each team memeber which I think Michael wrote it down

I also added the user story of the seller. I changed the format of some of them also.